+++ src/components/window/menubar.tsx
@@ -1,6 +1,8 @@
 import { Menubar, MenubarContent, MenubarGroup, MenubarItem, MenubarMenu, MenubarSeparator, MenubarShortcut, MenubarSub, MenubarSubContent, MenubarSubTrigger, MenubarTrigger } from "../ui/menubar";
 import { exit } from '@tauri-apps/plugin-process';
 
+// TODO: make each item alongside the logo dropdown working once the mockup is finished
+
 export default function MenuBar(){
      return (
           <Menubar className="h-full border-0 bg-transparent shadow-none rounded-none">
@@ -51,11 +53,18 @@ export default function MenuBar(){
                               <MenubarItem>Replace Translation <MenubarShortcut>Ctrl+R</MenubarShortcut></MenubarItem>
                               <MenubarItem>Batch Rename <MenubarShortcut>Ctrl+Shift+R</MenubarShortcut></MenubarItem>
                          </MenubarGroup>
+                         <MenubarSeparator/>
+                         <MenubarGroup>
+                              <MenubarItem>Go to prop name</MenubarItem>
+                              <MenubarItem>Select untranslated</MenubarItem>
+                              <MenubarItem>Compare diff</MenubarItem>
+                         </MenubarGroup>
                          <MenubarSeparator />
                          <MenubarGroup>
                               <MenubarItem>Cut <MenubarShortcut>Ctrl+X</MenubarShortcut></MenubarItem>
                               <MenubarItem>Copy <MenubarShortcut>Ctrl+C</MenubarShortcut></MenubarItem>
                               <MenubarItem>Paste <MenubarShortcut>Ctrl+V</MenubarShortcut></MenubarItem>
+                              <MenubarItem>Select All <MenubarShortcut>Ctrl+A</MenubarShortcut></MenubarItem>
                          </MenubarGroup>
                     </MenubarContent>
                </MenubarMenu>
@@ -70,9 +79,24 @@ export default function MenuBar(){
                <MenubarMenu>
                     <MenubarTrigger className="tracking-tight">Tools</MenubarTrigger>
                     <MenubarContent>
-                         <MenubarItem>Auto-translate</MenubarItem>
-                         <MenubarItem>Validate Keys</MenubarItem>
-                         <MenubarItem>Remove Unused Keys</MenubarItem>
+                         <MenubarSub>
+                              <MenubarSubTrigger>Translate using</MenubarSubTrigger>
+                              <MenubarSubContent>
+                                   <MenubarItem>Google Translate</MenubarItem>
+                                   <MenubarItem>Gemini</MenubarItem>
+                                   <MenubarItem>LibreTranslate</MenubarItem>
+                                   <MenubarItem>Llama AI</MenubarItem>
+                              </MenubarSubContent>
+                         </MenubarSub>
+                         <MenubarSub>
+                              <MenubarSubTrigger>Validation</MenubarSubTrigger>
+                              <MenubarSubContent>
+                                   <MenubarItem>Validate Keys</MenubarItem>
+                                   <MenubarItem>Remove Unused Keys</MenubarItem>
+                                   <MenubarItem>Spell check (using Hunspell)</MenubarItem>
+                              </MenubarSubContent>
+                         </MenubarSub>
+                         <MenubarItem>Transliterate Script</MenubarItem>
                     </MenubarContent>
                </MenubarMenu>
           </Menubar>